refactor: improve CLI structure
