+++ cmd/extensions/list/cmd.go
@@ -1,4 +1,4 @@
-package cmd
+package list
 
 import (
 	"fmt"
@@ -6,9 +6,9 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// InstalledCmd represents the installed command
-var InstalledCmd = &cobra.Command{
-	Use:   "installed",
+// ListCmd represents the list command
+var ListCmd = &cobra.Command{
+	Use:   "list",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -17,20 +17,18 @@ Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("installed called")
+		fmt.Println("list called")
 	},
 }
 
 func init() {
-	ExtensionsCmd.AddCommand(InstalledCmd)
-
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// InstalledCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// ListCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// InstalledCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// ListCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
